pyproject.toml
[project]
name = "analista-di-bilancio"
version = "0.1.0"
description = "Sistema agentivo per analisi automatizzata di bilanci aziendali italiani"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pdfplumber>=0.11.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
